feat: move Flex to f36-layout
