+++ packages/components/layout/src/index.ts
@@ -1,1 +1,2 @@
 export * from './Grid';
+export * from './Flex';
